# backend/services/product_service/docs/swagger.yaml
basePath: /
definitions:
...
  product_service_internal_core_port_service_dto.CategoryRes:
    properties:
      children:
        items:
          $ref: '#/definitions/product_service_internal_core_port_service_dto.CategoryRes'
        type: array
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      parent_id:
        type: integer
      slug:
        type: string
      updated_at:
        type: string
    type: object
...
  product_service_internal_core_port_service_dto.CreateCategoryReq:
    properties:
      description:
        maxLength: 255
        type: string
      is_active:
        type: boolean
      name:
        maxLength: 100
        minLength: 1
        type: string
      parent_id:
        type: integer
      slug:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - name
    - slug
    type: object
...
  product_service_internal_core_port_service_dto.SetCategoryActiveReq:
    properties:
      category_id:
        description: ถูก populate จาก URL param โดย Handler
        type: integer
      is_active:
        type: boolean
    type: object
...
  product_service_internal_core_port_service_dto.UpdateCategoryReq:
    properties:
      category_id:
        type: integer
      description:
        maxLength: 255
        type: string
      is_active:
        type: boolean
      name:
        maxLength: 100
        minLength: 1
        type: string
      parent_id:
        type: integer
      slug:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - category_id
    - name
    - slug
    type: object
...
host: localhost:3002
info:
  contact:
    email: your.email@example.com
    name: API Support
  description: This is the Product Service API using Hexagonal Architecture & CQRS.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Product Service API
  version: "1.0"
paths:
...
  /categories:
    get:
      description: Returns all root categories with their children (tree structure).
      produces:
      - application/json
      responses:
        "200":
          description: Category tree
          schema:
            items:
              $ref: '#/definitions/product_service_internal_core_port_service_dto.CategoryRes'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: List all categories
      tags:
      - Categories (Query)
  /categories/{id}:
    get:
      description: Returns a single category with its children.
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Category detail
          schema:
            $ref: '#/definitions/product_service_internal_core_port_service_dto.CategoryRes'
        "404":
          description: Category not found
          schema:
            additionalProperties: true
            type: object
      summary: Get category by ID
      tags:
      - Categories (Query)
  /categories/admin:
    post:
      consumes:
      - application/json
      description: Creates a new product category. Supports nested categories via
        parent_id.
      parameters:
      - description: Category data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/product_service_internal_core_port_service_dto.CreateCategoryReq'
      produces:
      - application/json
      responses:
        "201":
          description: Category created successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new category (Admin)
      tags:
      - Categories (Command)
  /categories/admin/{id}:
    delete:
      description: Soft-deletes a category by ID.
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Category deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Category not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a category (Admin)
      tags:
      - Categories (Command)
    put:
      consumes:
      - application/json
      description: Updates name, slug, description, is_active, and parent_id of a
        category.
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated category data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/product_service_internal_core_port_service_dto.UpdateCategoryReq'
      produces:
      - application/json
      responses:
        "200":
          description: Category updated successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Category not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a category (Admin)
      tags:
      - Categories (Command)
  /categories/admin/{id}/active:
    patch:
      consumes:
      - application/json
      description: Toggle category visibility. is_active=false hides it from the storefront.
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Active status
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/product_service_internal_core_port_service_dto.SetCategoryActiveReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Activate or deactivate a category (Admin)
      tags:
      - Categories (Command)
...
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer " followed by your token. Example: "Bearer eyJhbGci..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"